Mejora preparada para cambiar matsnackbar, dialog de confirmación mejora y dialog con tabla de documentos anexos

+++ front/src/app/kanban/components/kanban-table/kanban-table.component.ts
@@ -6,6 +6,8 @@ import {
   moveItemInArray,
   transferArrayItem,
 } from '@angular/cdk/drag-drop';
+import { MatDialog } from '@angular/material/dialog';
+import { DocumentsDialogComponent } from '../documents-dialog/documents-dialog.component';
 
 @Component({
   selector: 'app-kanban-table',
@@ -18,7 +20,7 @@ export class KanbanTableComponent {
   todo: Task[] = [];
   done: Task[] = [];
 
-  constructor(private kanbanService: KanbanService) { }
+  constructor(private kanbanService: KanbanService, private dialog: MatDialog) { }
 
   ngOnInit() {
     const userId = Number(localStorage.getItem('user_id'));
@@ -50,14 +52,16 @@ export class KanbanTableComponent {
             event.previousIndex,
             event.currentIndex
           );
-          console.log(`Tarea actualizada con éxito a estado ${state}`) // cambiar por matsnackbar
+          alert(`Tarea actualizada con éxito a estado ${state}`) // cambiar por matsnackbar
         },
-        error: (err) => console.error('Error actualizando la tarea:', err) // cambiar por matsnackbar
+        error: (err) => alert('Error actualizando la tarea: ' + err) // cambiar por matsnackbar
       });
     }
   }
 
   deleteTask(id_task: number) {
+    const confirmDelete = confirm("¿Estás seguro de que quieres eliminar esta tarea?"); // cambiar por dialog de confirmación
+    if (!confirmDelete) return; // cambiar por dialog de confirmación
     const userId = Number(localStorage.getItem('user_id'));
 
     this.kanbanService.deleteTask(id_task, userId).subscribe({
@@ -67,6 +71,15 @@ export class KanbanTableComponent {
         console.log('Tarea eliminada con éxito.'); // cambiar por matsnackbar
       },
       error: (err) => console.error('Error al intentar eliminar la tarea.', err) // cambiar por matsnackbar
+    });
+  }
+
+  openDialog() {
+    this.dialog.open(DocumentsDialogComponent, {
+      width: 'auto',
+      enterAnimationDuration: '300ms',
+      exitAnimationDuration: '300ms',
+      data: { title: 'Documentos anexos', button: 'Cerrar', message: 'Esta es la documentación anexa que tienes que completar para rellenar la tarea.' }
     });
   }
 }
